REPLAT-671: Design Review applied
