+++ packages/article-label/style/index.android.js
@@ -1,0 +1,11 @@
+import sharedStyles from "./shared";
+
+const styles = {
+  title: {
+    ...sharedStyles.title,
+    paddingTop: 5,
+    marginBottom: -4
+  }
+}
+
+export default styles;
