+++ packages/article-label/style/index.js
@@ -1,0 +1,9 @@
+import sharedStyles from "./shared";
+
+const styles = {
+  title: {
+    ...sharedStyles.title
+  }
+}
+
+export default styles;
