+++ packages/article-label/style/shared.js
@@ -1,0 +1,11 @@
+const styles = {
+  title: {
+    fontFamily: "GillSansMTStd-Medium",
+    fontSize: 12,
+    lineHeight: 12,
+    fontWeight: "400",
+    letterSpacing: 1.2
+  }
+};
+
+export default styles;
